requirements.txt
sounddevice
scipy
python-dotenv
vosk
elevenlabs
google-genai
pydub
simpleaudio
